requirements.txt
pandas
numpy
tensorflow==2.17.1
dvc
mlflow==2.2.2
notebook
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors
gdown

-e .
